==> 2015/02_Shapefiles_2015.Rmd <==
---
title: "Shapefiles 2015"
author: "Diana Villasana Ocampo"
output:
   html_document:
      highlight: tango
      theme: flatly
      toc: yes
      toc_depth: 2
      toc_float:
        collapsed: yes
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = TRUE, message = FALSE, warning = FALSE, cache = FALSE, 
                      cache.lazy = FALSE, eval = FALSE, class.source = "fold-show")
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
options(digits = 2, encoding = "UTF8")
```   
 

```{r, echo=FALSE}
rm(list = ls())
```

```{r, echo=FALSE}
setwd(here::here())
```

**Libraries** 

```{r}
require(dplyr)
require(forcats)
require(rgdal)
require(tibble)
require(sf)
require(sp)
require(stringr)
```


## Índice de marginación 2015

**Bases de datos***
- Nivel estatal
- Nivel municipal
- Nivel localidad
- Nivel AGEB 


```{r}
load(paste0(here::here(), "/2015/Output/IMM_2015.RData"))
```


## Shapefiles

**Shapefile estados 2015**

```{r}
shape_estados <- readOGR(dsn ="D:/MGN/MGN Junio 2016/conjunto_de_datos", 
                          layer = "areas_geoestadisticas_estatales",
                            encoding = "UTF-8",
                             use_iconv = TRUE)

shape_estados <- spTransform(shape_estados, CRSobj = CRS('+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'))
```


**Shapefile municipios 2015**

```{r}
shape_municipios <- readOGR(dsn ="D:/MGN/MGN Junio 2016/conjunto_de_datos", 
                             layer = "areas_geoestadisticas_municipales",
                              encoding = "UTF-8",
                               use_iconv = TRUE)

shape_municipios <- spTransform(shape_municipios, CRSobj = CRS('+proj=longlat +ellps=WGS84 +datum=WGS84 +no_defs'))
```


## ShapefilesDataFrame

**Se insertan los datos del índice de marginación y se seccionan por grado de marginación**

### Municipios

```{r}
shape_municipios@data <- shape_municipios@data %>%
                          mutate(CVE_MUN = paste0(.$CVE_ENT, .$CVE_MUN)) %>%
                           select(CVE_MUN)

shape_municipios <- shape_municipios %>%
                     sp::spChFIDs(., str_pad(shape_municipios@data$CVE_MUN, 5, "left", pad = "0"))
```


#### Capas de municipios

- `Muy alto` ~ `1` | 275
- `Alto` ~ `2` | 517
- `Medio` ~ `3` | 617 
- `Bajo` ~ `4` | 601
- `Muy bajo` ~ `5` | 447


```{r}
capa_municipios_1 <- shape_municipios %>%
                      spdplyr:::inner_join.Spatial(., IMM_2015 %>% 
                                                       select(CVE_MUN, GM_2015) %>%
                                                        filter(GM_2015 == "Muy alto"),
                                                          by = c("CVE_MUN")) %>%
                       select(CVE_MUN)
capa_municipios_2 <- shape_municipios %>%
                      spdplyr:::inner_join.Spatial(., IMM_2015 %>% 
                                                       select(CVE_MUN, GM_2015) %>%
                                                        filter(GM_2015 == "Alto"),
                                                          by = c("CVE_MUN"))  %>%
                       select(CVE_MUN)

capa_municipios_3 <- shape_municipios %>%
                      spdplyr:::inner_join.Spatial(., IMM_2015 %>% 
                                                       select(CVE_MUN, GM_2015) %>%
                                                        filter(GM_2015 == "Medio"),
                                                          by = c("CVE_MUN"))  %>%
                       select(CVE_MUN)

capa_municipios_4 <- shape_municipios %>%
                      spdplyr:::inner_join.Spatial(., IMM_2015 %>% 
                                                       select(CVE_MUN, GM_2015) %>%
                                                        filter(GM_2015 == "Bajo"),
                                                          by = c("CVE_MUN"))  %>%
                       select(CVE_MUN)
capa_municipios_5 <- shape_municipios %>%
                      spdplyr:::inner_join.Spatial(., IMM_2015 %>% 
                                                       select(CVE_MUN, GM_2015) %>%
                                                        filter(GM_2015 == "Muy bajo"),
                                                          by = c("CVE_MUN"))  %>%
                       select(CVE_MUN)
```

**Se guardan los shapefiles en un archivo `RDS`**

```{r}
saveRDS(shape_estados, paste0(here::here(), "/2015/Output/Shapes/shape_estados.rds"))
saveRDS(capa_municipios_1, paste0(here::here(), "/2015/Output/Shapes/shape_municipios_1.rds"))
saveRDS(capa_municipios_2, paste0(here::here(), "/2015/Output/Shapes/shape_municipios_2.rds"))
saveRDS(capa_municipios_3, paste0(here::here(), "/2015/Output/Shapes/shape_municipios_3.rds"))
saveRDS(capa_municipios_4, paste0(here::here(), "/2015/Output/Shapes/shape_municipios_4.rds"))
saveRDS(capa_municipios_5, paste0(here::here(), "/2015/Output/Shapes/shape_municipios_5.rds"))
```


#### Geojson 

```{r}
require(geojsonio)
capa_estados_json <- geojsonio::geojson_json(shape_estados, geometry = "polygon")
capa_municipios_1_json <- geojsonio::geojson_json(capa_municipios_1 %>%
                                                   sp::spChFIDs(., str_pad(.@data$CVE_MUN, 5, "left", pad = "0")), 
                                                    geometry = "polygon")
capa_municipios_2_json <- geojsonio::geojson_json(capa_municipios_2 %>%
                                                   sp::spChFIDs(., str_pad(.@data$CVE_MUN, 5, "left", pad = "0")), 
                                                    geometry = "polygon")
capa_municipios_3_json <- geojsonio::geojson_json(capa_municipios_3 %>%
                                                   sp::spChFIDs(., str_pad(.@data$CVE_MUN, 5, "left", pad = "0")), 
                                                    geometry = "polygon")
capa_municipios_4_json <- geojsonio::geojson_json(capa_municipios_4 %>%
                                                   sp::spChFIDs(., str_pad(.@data$CVE_MUN, 5, "left", pad = "0")), 
                                                    geometry = "polygon")
capa_municipios_5_json <- geojsonio::geojson_json(capa_municipios_5 %>%
                                                   sp::spChFIDs(., str_pad(.@data$CVE_MUN, 5, "left", pad = "0")), 
                                                    geometry = "polygon")
```


### Reducción de los mapas 

```{r}
require(rmapshaper)
# Se reduce el tamaño y se guarda como archivo geojason 
capa_estados_mapshaper <- rmapshaper::ms_simplify(capa_estados_json, keep = 0.3, weighting = 0, keep_shapes = TRUE)
capa_municipios_1_mapshaper <- rmapshaper::ms_simplify(capa_municipios_1_json, keep = 0.2, weighting = 0, keep_shapes = TRUE)
capa_municipios_2_mapshaper <- rmapshaper::ms_simplify(capa_municipios_2_json, keep = 0.2, weighting = 0, keep_shapes = TRUE)
capa_municipios_3_mapshaper <- rmapshaper::ms_simplify(capa_municipios_3_json, keep = 0.2, weighting = 0, keep_shapes = TRUE)
capa_municipios_4_mapshaper <- rmapshaper::ms_simplify(capa_municipios_4_json, keep = 0.2, weighting = 0, keep_shapes = TRUE)
capa_municipios_5_mapshaper <- rmapshaper::ms_simplify(capa_municipios_5_json, keep = 0.2, weighting = 0, keep_shapes = TRUE)
```


```{r}
# Se guarda en un archivo geojson 
geojsonio::geojson_write(capa_estados_mapshaper, file = paste0(here::here(), "/2015/Output/Shapes/estados_json.geojson"))
geojsonio::geojson_write(capa_municipios_1_mapshaper, file = paste0(here::here(), "/2015/Output/Shapes/municipios_1_json.geojson"))
geojsonio::geojson_write(capa_municipios_2_mapshaper, file = paste0(here::here(), "/2015/Output/Shapes/municipios_2_json.geojson"))
geojsonio::geojson_write(capa_municipios_3_mapshaper, file = paste0(here::here(), "/2015/Output/Shapes/municipios_3_json.geojson"))
geojsonio::geojson_write(capa_municipios_4_mapshaper, file = paste0(here::here(), "/2015/Output/Shapes/municipios_4_json.geojson"))
geojsonio::geojson_write(capa_municipios_5_mapshaper, file = paste0(here::here(), "/2015/Output/Shapes/municipios_5_json.geojson"))
```

```{r}
# Se guarda como objeto
saveRDS(capa_estados_mapshaper, paste0(here::here(), "/2015/Output/Shapes/capa_estados_mapshaper.rds"))
saveRDS(capa_municipios_1_mapshaper, paste0(here::here(), "/2015/Output/Shapes/capa_municipios_1_mapshaper.rds"))
saveRDS(capa_municipios_2_mapshaper, paste0(here::here(), "/2015/Output/Shapes/capa_municipios_2_mapshaper.rds"))
saveRDS(capa_municipios_3_mapshaper, paste0(here::here(), "/2015/Output/Shapes/capa_municipios_3_mapshaper.rds"))
saveRDS(capa_municipios_4_mapshaper, paste0(here::here(), "/2015/Output/Shapes/capa_municipios_4_mapshaper.rds"))
saveRDS(capa_municipios_5_mapshaper, paste0(here::here(), "/2015/Output/Shapes/capa_municipios_5_mapshaper.rds"))
```



```{r}
## Shapefiles
shape_estados <- geojsonio::geojson_read(paste0(here::here(), "/2015/Output/Shapes/estados_json.geojson"), what = "sp")
```
### Topojson

```{r}
require(geojsonio)
capa_estados_topojson <- capa_estados_mapshaper %>% geojsonio::geo2topo()
```

```{r}
# Se guarda como objeto
saveRDS(capa_estados_topojson, paste0(here::here(), "/2015/Output/Shapes/capa_estados_topojson.rds"))
```
